Affected Issue(s): https://www.notion.so/Back-End-add-criteria-is_consented_whatsapp-to-every-query-automated-message-825daec9117a40e99b541952fea919a0 https://www.notion.so/Back-End-Disable-non-preggo-intro-message-9611833105fe414f873148e77cd390d8
What this commit has achieved:
1. Disabled non-preggo automated intro message
2. Added new field `isConsentedWhatsapp` to `AncRegister` domain
